fix: generalize cleanup to apply to other installer
